+++ CHANGELOG.md
@@ -44,6 +44,10 @@ Section Order:
 > Please make sure to update your Alliance Auth instance **before** you install this
 > version, otherwise an update to Alliance Auth will be pulled in unsupervised.
 
+### Changed
+
+- DataTables are now responsive
+
 ### Removed
 
 - Support for Alliance Auth v4